Simplex noise on sphere + density arg

+++ include/sphere_mesh.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <glm/glm.hpp>
 #include "type.hpp"
+#include <glm/glm.hpp>
 #include <ostream>
 #include <vector>
 
@@ -10,11 +10,15 @@ class SphereMesh {
   friend std::ostream &operator<<(std::ostream &o, const SphereMesh &mesh);
 
 public:
-  SphereMesh(int radius_, int density);
+  SphereMesh(float radius_, int density);
   std::ostream &operator<<(std::ostream &o);
 
-  std::vector<type::Vertex>& getVertices()   { return vertices; }
-  std::vector<type::Face>& getFaces()      { return faces; }
+  std::vector<type::Vertex> &getVertices() {
+    return vertices;
+  }
+  std::vector<type::Face> &getFaces() {
+    return faces;
+  }
 
 private:
   void Generate(int density);
@@ -23,8 +27,8 @@ private:
   void addQuad(int a, int b, int c, int d);
   void addQuadAlt(int a, int b, int c, int d);
 
-  int radius;
-  
+  float radius;
+
   std::vector<type::Vertex> vertices;
   std::vector<type::Face> faces;
 };
